# api/openapi.yaml
openapi: 3.0.3
info:
  description: API to retrieve and search GPS Walking Routes
  title: GPS Routes API
  version: 0.0.1
servers:
- url: https://api.destructuring-bind.org
paths:
  /v1/gps-routes/search:
    post:
      description: Returns multiple matching summary results in JSON format.
      operationId: search
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
          description: Successful response
        "400":
          description: "Invalid request, e.g., missing or invalid request body attributes."
        "500":
          description: Internal server error
      summary: Search for routes according to various criteria
      tags:
      - routes
  /v1/gps-routes/{objectID}:
    get:
      description: Returns data in JSON format.
      operationId: fetchRecord
      parameters:
      - description: The object ID exactly describing the GPS route (typically this
          is the MD5 hash of the route reference)
        explode: false
        in: path
        name: objectID
        required: true
        schema:
          example: 4a2bb897585c84113beeb3f2691d9042
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteMetadata'
          description: Successful response
        "400":
          description: "Invalid request, e.g., missing or invalid path parameter."
        "404":
          description: "Not found, e.g., no such record for the specified object ID."
        "500":
          description: Internal server error
      summary: Retrieve metadata for the specific walking route
      tags:
      - routes
components:
  schemas:
    RouteSummary:
      example:
        ref: daneshill-lakes-walking-route
        description: "This easy walk takes you around the lovely Daneshill Lakes Nature\
          \ Reserve near Retford. There's wetlands, woodland, grassland, flowering\
          \ plants and lots of wildlife to look out for on the site. Look out for\
          \ grebes, swans, willow warbler, whitethroat, and blackcap on your walk.\
          \ There's a free on-site car park from which you can start the walk. To\
          \ the south there's the more popular open water section, while to the north\
          \ there's some peaceful woodland trails to enjoy. The reserve is located\
          \ just to the west of Idle Valley Nature Reserve. The expansive reserve\
          \ covers 450 hectares with hundreds of species of wetland birds to look\
          \ out for. To the south of the site you can pick up the Chesterfield Canal\
          \ to continue your waterside walking. The towpath can be followed into Retford\
          \ or towards Ranby."
        location:
          coordinates:
          - -0.9960077399
          - 53.3712125339
          type: Point
        title: Daneshill Lakes - Retford Walking Route
        objectID: 0a0cafcd50f17284dc578271968f8341
        headline_image_url: https://upload.wikimedia.org/wikipedia/commons/e/e5/Daneshill_Lakes_-_Car_Park_Entrance_-_geograph.org.uk_-_674042.jpg
      properties:
        objectID:
          description: The object ID identifies the GPS route (typically this is the
            MD5 hash of the route reference)
          example: 0a0cafcd50f17284dc578271968f8341
          type: string
        ref:
          description: A human friendly unique identifier for the route.
          example: daneshill-lakes-walking-route
          type: string
        title:
          description: The route title
          example: Daneshill Lakes - Retford Walking Route
          type: string
        description:
          description: Typically a long passage of text describing aspects of the
            route
          example: "This easy walk takes you around the lovely Daneshill Lakes Nature\
            \ Reserve near Retford. There's wetlands, woodland, grassland, flowering\
            \ plants and lots of wildlife to look out for on the site. Look out for\
            \ grebes, swans, willow warbler, whitethroat, and blackcap on your walk.\
            \ There's a free on-site car park from which you can start the walk. To\
            \ the south there's the more popular open water section, while to the\
            \ north there's some peaceful woodland trails to enjoy. The reserve is\
            \ located just to the west of Idle Valley Nature Reserve. The expansive\
            \ reserve covers 450 hectares with hundreds of species of wetland birds\
            \ to look out for. To the south of the site you can pick up the Chesterfield\
            \ Canal to continue your waterside walking. The towpath can be followed\
            \ into Retford or towards Ranby."
          type: string
        headline_image_url:
          description: The main image URL associated with the route
          example: https://upload.wikimedia.org/wikipedia/commons/e/e5/Daneshill_Lakes_-_Car_Park_Entrance_-_geograph.org.uk_-_674042.jpg
          format: url
          type: string
        location:
          $ref: '#/components/schemas/GeoJSON'
      required:
      - description
      - headline_image_url
      - objectID
      - ref
      - title
      type: object
    RouteMetadata:
      allOf:
      - $ref: '#/components/schemas/RouteSummary'
      - properties:
          created_at:
            description: The time the route was generated.
            example: 2024-12-18T16:35:30.796469Z
            format: date-time
            type: string
          nearby:
            items:
              $ref: '#/components/schemas/Nearby'
            type: array
          gpx_url:
            description: URL pointing to the associated GPX route data
            format: url
            type: string
          images:
            items:
              $ref: '#/components/schemas/Image'
            type: array
          distance_km:
            description: A rough distance indicating the length of the route
            example: 1.7
            type: float64
          video_url:
            description: A link to a (YouTube) video associated with the route
            example: https://www.youtube.com/embed/U8pyBfqKKhE
            format: url
            type: string
          postcode:
            description: "Where exists, a nearby postcode"
            type: string
            example: null
          district:
            description: The district in which the route starts
            type: string
            example: null
          county:
            description: The county in which the route starts
            example: Nottinghamshire
            type: string
          region:
            description: The region in which the route starts
            example: East Midlands
            type: string
          country:
            description: The country in which the route starts
            example: England
            type: string
          estimated_duration:
            description: The estimated time to complete the route
            example: 2-3 hours
            type: string
          difficulty:
            description: The difficulty of the route
            example: easy
            type: string
          terrain:
            description: The terrain of the route
            example: ["flat", "river", "woodland"]
            type: array
          points_of_interest:
            description: Points of interest along the route
            example: ["castle", "pub", "church"]
            type: array
          facilities:
            description: Facilities available along the route
            example: ["cafe", "toilets", "parking"]
            type: array
          route_type:
            description: The type of route
            example: "circular"
            type: string
          activities:
            description: Activities available along the route
            example: ["birdwatching", "cycling", "fishing"]
            type: array
        required:
        - created_at
        type: object
      example:
        country: England
        images:
        - src: https://upload.wikimedia.org/wikipedia/commons/thumb/2/24/Footpath_and_meadow_at_Daneshill_Lakes_reserve_-_geograph.org.uk_-_1948823.jpg/512px-Footpath_and_meadow_at_Daneshill_Lakes_reserve_-_geograph.org.uk_-_1948823.jpg
          caption: Footpath and meadow
          title: "Footpath and meadow at Daneshill Lakes reserve by Andrew Hill, CC\
            \ BY-SA 2.0 <https://creativecommons.org/licenses/by-sa/2.0>, via Wikimedia\
            \ <https://creativecommons.org/licenses/by-sa/2.0"
        - src: https://upload.wikimedia.org/wikipedia/commons/thumb/2/24/Footpath_and_meadow_at_Daneshill_Lakes_reserve_-_geograph.org.uk_-_1948823.jpg/512px-Footpath_and_meadow_at_Daneshill_Lakes_reserve_-_geograph.org.uk_-_1948823.jpg
          caption: Footpath and meadow
          title: "Footpath and meadow at Daneshill Lakes reserve by Andrew Hill, CC\
            \ BY-SA 2.0 <https://creativecommons.org/licenses/by-sa/2.0>, via Wikimedia\
            \ <https://creativecommons.org/licenses/by-sa/2.0"
        distance_km: 1.7
        postcode: postcode
        county: Nottinghamshire
        description: "This easy walk takes you around the lovely Daneshill Lakes Nature\
          \ Reserve near Retford. There's wetlands, woodland, grassland, flowering\
          \ plants and lots of wildlife to look out for on the site. Look out for\
          \ grebes, swans, willow warbler, whitethroat, and blackcap on your walk.\
          \ There's a free on-site car park from which you can start the walk. To\
          \ the south there's the more popular open water section, while to the north\
          \ there's some peaceful woodland trails to enjoy. The reserve is located\
          \ just to the west of Idle Valley Nature Reserve. The expansive reserve\
          \ covers 450 hectares with hundreds of species of wetland birds to look\
          \ out for. To the south of the site you can pick up the Chesterfield Canal\
          \ to continue your waterside walking. The towpath can be followed into Retford\
          \ or towards Ranby."
        created_at: 2024-12-18T16:35:30.796469Z
        title: Daneshill Lakes - Retford Walking Route
        nearby:
        - ref: chesterfield-canal-walking-route
          description: Chesterfield Canal
          objectID: 82d4977c39ae50e61f209ab628ffe8ae
        - ref: chesterfield-canal-walking-route
          description: Chesterfield Canal
          objectID: 82d4977c39ae50e61f209ab628ffe8ae
        headline_image_url: https://upload.wikimedia.org/wikipedia/commons/e/e5/Daneshill_Lakes_-_Car_Park_Entrance_-_geograph.org.uk_-_674042.jpg
        ref: daneshill-lakes-walking-route
        video_url: https://www.youtube.com/embed/U8pyBfqKKhE
        gpx_url: http://example.com/aeiou
        district: district
        location:
          coordinates:
          - -0.9960077399
          - 53.3712125339
          type: Point
        region: East Midlands
        objectID: 0a0cafcd50f17284dc578271968f8341
        estimated_duration: 1.5 - 2.5 hours
        difficulty: easy
        terrain: 
          - woodland
          - wetland
          - grassland
        points_of_interest:
          - wildlife viewing
          - nature trails
        facilities:
          - car park
          - canal access
        route_type: loop,
        activities:
          - walking
    Nearby:
      example:
        ref: chesterfield-canal-walking-route
        description: Chesterfield Canal
        objectID: 82d4977c39ae50e61f209ab628ffe8ae
      properties:
        objectID:
          description: The object ID identifies the GPS route (typically this is the
            MD5 hash of the route reference)
          example: 82d4977c39ae50e61f209ab628ffe8ae
          type: string
        ref:
          description: A human friendly unique identifier for the route.
          example: chesterfield-canal-walking-route
          type: string
        description:
          description: The route title
          example: Chesterfield Canal
          type: string
      required:
      - description
      - objectID
      - ref
      type: object
    Image:
      example:
        src: https://upload.wikimedia.org/wikipedia/commons/thumb/2/24/Footpath_and_meadow_at_Daneshill_Lakes_reserve_-_geograph.org.uk_-_1948823.jpg/512px-Footpath_and_meadow_at_Daneshill_Lakes_reserve_-_geograph.org.uk_-_1948823.jpg
        caption: Footpath and meadow
        title: "Footpath and meadow at Daneshill Lakes reserve by Andrew Hill, CC\
          \ BY-SA 2.0 <https://creativecommons.org/licenses/by-sa/2.0>, via Wikimedia\
          \ <https://creativecommons.org/licenses/by-sa/2.0"
      properties:
        src:
          description: A URL referencing an image
          example: https://upload.wikimedia.org/wikipedia/commons/thumb/2/24/Footpath_and_meadow_at_Daneshill_Lakes_reserve_-_geograph.org.uk_-_1948823.jpg/512px-Footpath_and_meadow_at_Daneshill_Lakes_reserve_-_geograph.org.uk_-_1948823.jpg
          format: url
          type: string
        title:
          description: "Description, typically containing some attribution for the\
            \ image"
          example: "Footpath and meadow at Daneshill Lakes reserve by Andrew Hill,\
            \ CC BY-SA 2.0 <https://creativecommons.org/licenses/by-sa/2.0>, via Wikimedia\
            \ <https://creativecommons.org/licenses/by-sa/2.0"
          type: string
        caption:
          description: Appropriate image caption
          example: Footpath and meadow
          type: string
      required:
      - src
      type: object
    GeoJSON:
      description: A GeoJSON object
      example:
        coordinates:
        - -0.9960077399
        - 53.3712125339
        type: Point
      properties:
        type:
          description: GeoJSON type
          example: Point
          type: string
        coordinates:
          description: "Longitude, Latitude"
          example:
          - -0.9960077399
          - 53.3712125339
          items:
            type: float64
          type: array
      required:
      - coordinates
      - type
      type: object
    SearchRequest:
      example:
        boundingBox:
        - -2.945623397827149
        - 54.43847718997843
        - -2.9769945144653325
        - 54.41890499817318
        offset: 0
        query: coffee -tea shop
        limit: 6
        facets:
          key:
          - facets
          - facets
      properties:
        query:
          description: Search phrase
          example: coffee -tea shop
          type: string
        offset:
          type: integer
        limit:
          type: integer
        boundingBox:
          description: "the bounding box with bottom-left lng/lat, followed by top-right\
            \ lng/lat"
          example:
          - -2.945623397827149
          - 54.43847718997843
          - -2.9769945144653325
          - 54.41890499817318
          items:
            type: float64
          type: array
        facets:
          additionalProperties:
            items:
              type: string
            type: array
          description: Filtering by facet values
          type: object
      required:
      - query
      type: object
    SearchResults:
      example:
        hits:
        - ref: daneshill-lakes-walking-route
          description: "This easy walk takes you around the lovely Daneshill Lakes\
            \ Nature Reserve near Retford. There's wetlands, woodland, grassland,\
            \ flowering plants and lots of wildlife to look out for on the site. Look\
            \ out for grebes, swans, willow warbler, whitethroat, and blackcap on\
            \ your walk. There's a free on-site car park from which you can start\
            \ the walk. To the south there's the more popular open water section,\
            \ while to the north there's some peaceful woodland trails to enjoy. The\
            \ reserve is located just to the west of Idle Valley Nature Reserve. The\
            \ expansive reserve covers 450 hectares with hundreds of species of wetland\
            \ birds to look out for. To the south of the site you can pick up the\
            \ Chesterfield Canal to continue your waterside walking. The towpath can\
            \ be followed into Retford or towards Ranby."
          location:
            coordinates:
            - -0.9960077399
            - 53.3712125339
            type: Point
          title: Daneshill Lakes - Retford Walking Route
          objectID: 0a0cafcd50f17284dc578271968f8341
          headline_image_url: https://upload.wikimedia.org/wikipedia/commons/e/e5/Daneshill_Lakes_-_Car_Park_Entrance_-_geograph.org.uk_-_674042.jpg
        - ref: daneshill-lakes-walking-route
          description: "This easy walk takes you around the lovely Daneshill Lakes\
            \ Nature Reserve near Retford. There's wetlands, woodland, grassland,\
            \ flowering plants and lots of wildlife to look out for on the site. Look\
            \ out for grebes, swans, willow warbler, whitethroat, and blackcap on\
            \ your walk. There's a free on-site car park from which you can start\
            \ the walk. To the south there's the more popular open water section,\
            \ while to the north there's some peaceful woodland trails to enjoy. The\
            \ reserve is located just to the west of Idle Valley Nature Reserve. The\
            \ expansive reserve covers 450 hectares with hundreds of species of wetland\
            \ birds to look out for. To the south of the site you can pick up the\
            \ Chesterfield Canal to continue your waterside walking. The towpath can\
            \ be followed into Retford or towards Ranby."
          location:
            coordinates:
            - -0.9960077399
            - 53.3712125339
            type: Point
          title: Daneshill Lakes - Retford Walking Route
          objectID: 0a0cafcd50f17284dc578271968f8341
          headline_image_url: https://upload.wikimedia.org/wikipedia/commons/e/e5/Daneshill_Lakes_-_Car_Park_Entrance_-_geograph.org.uk_-_674042.jpg
        total: ""
        facets:
          key:
            key: ""
      properties:
        total:
          type: int64
        hits:
          items:
            $ref: '#/components/schemas/RouteSummary'
          type: array
        facets:
          additionalProperties:
            additionalProperties:
              type: int64
            type: object
          type: object
      required:
      - hits
      - total
      type: object