# routes/api_routes.go
/*
 * GPS Routes API
 *
 * API to retrieve and search GPS Walking Routes
 *
 * API version: 0.0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package routes

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"

	model "github.com/rm-hull/gps-routes-api/go"
	svc "github.com/rm-hull/gps-routes-api/services"
)

type RoutesAPI struct {
	Service svc.RoutesService
}

// Get /v1/gps-routes/:objectID
// Retrieve metadata for the specific walking route
func (api *RoutesAPI) FetchRecord(c *gin.Context) {
	objectID := c.Param("objectID")
	if objectID == "" {
		c.JSON(http.StatusBadRequest, gin.H{"error": "objectID is required"})
		return
	}
	route, err := api.Service.GetRouteByID(objectID)
	if err != nil {
		c.Error(err) //nolint:errcheck
		return
	}

	if route == nil {
		c.JSON(http.StatusNotFound, gin.H{"message": "Not found"})
		return
	}

	c.JSON(http.StatusOK, route)
}

var VALID_FACET_FIELDS = []string{
	"country", "state", "region", "district", "county",
	"activities", "facilities", "points_of_interest",
	"route_type", "difficulty", "terrain", "estimated_duration",
}

func isValidFacetField(facetField string) bool {
	for _, field := range VALID_FACET_FIELDS {
		if field == facetField {
			return true
		}
	}
	return false
}

// Post /v1/gps-routes/search
// Search for routes according to various criteria
func (api *RoutesAPI) Search(c *gin.Context) {
	var payload model.SearchRequest
	err := c.ShouldBind(&payload)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "bad payload"})
		return
	}

	// If limit not specified, default to 10
	if payload.Limit == 0 {
		payload.Limit = 10
	}

	if payload.Limit < 0 || payload.Limit > 100 {
		c.JSON(http.StatusBadRequest, gin.H{"error": "limit is out of range - allowed values: 0..100"})
		return
	}

	if payload.Offset < 0 || payload.Offset > 100_000 {
		c.JSON(http.StatusBadRequest, gin.H{"error": "offset is out of range - allowed values: 0..100_000"})
		return
	}

	for facetField := range payload.Facets {
		if !isValidFacetField(facetField) {
			c.JSON(http.StatusBadRequest, gin.H{"error": fmt.Sprintf("invalid facet: %s", facetField)})
			return
		}
	}

	// TODO: sanitize the query string

	matches, err := api.Service.Search(&payload)
	if err != nil {
		c.Error(err) //nolint:errcheck
		return
	}

	c.JSON(http.StatusOK, matches)

}